+++ packages/simple_dart_code_gen/lib/simple_dart_code_gen.dart
@@ -1318,7 +1318,7 @@ final class ExprSwitch implements Expr {
                     ',\n',
               )
               .safeJoin() +
-          '}',
+          '})',
       ConstType.noConst,
     );
   }